requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
aiomoex==1.2.2
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
gspread==3.3.0
gspread-asyncio==1.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.3.25
httplib2==0.17.0
httptools==0.1.1
httpx==0.10.1
hyperframe==5.2.0
idna==2.9
jwt==1.0.0
motor==2.1.0
multidict==4.7.5
numpy==1.18.2
oauth2client==4.1.3
pandas==1.0.3
passlib==1.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.4
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rfc3986==1.3.2
rsa==4.0
sentry-sdk==0.14.3
six==1.14.0
sniffio==1.1.0
starlette==0.13.2
tenacity==6.1.0
ujson==2.0.3
urllib3==1.25.8
uvloop==0.14.0
websockets==8.1
yarl==1.4.2
databases
uvicorn
loguru
tortoise-orm
asyncpg
aiosqlite
aiosql
aiomysql
pydantic[email]==1.6.1
fastapi==0.54.1
python-slugify
email_validator
jinja2